Fix OperationStats.count type and reason code extraction regression
- count should be int not float (it tracks call counts)
- _extract_reason_codes_from_observation now handles missing code_system
  (produces text-only CodeableConcept), matching old create_codeable_concept behavior

+++ ccda_to_fhir/types.py
@@ -105,7 +105,7 @@ class DiagnosisRole(BaseModel, frozen=True):
 class OperationStats(BaseModel, frozen=True):
     """Performance statistics for a single profiled operation."""
 
-    count: float
+    count: int
     total: float
     avg: float
     min: float
